fix: exception thrown fixed
An exception was being thrown by the code when it detected no features 
in the training image. This was fixed by using an if statement that 
checked if features were detected in both the training and query image 
or not.
Code formatting was changed and new comments were added.
Readme file was updated.

Resolves: #3